% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bokeh_colorblind.r
\docType{data}
\name{bokeh_colorblind}
\alias{bokeh_colorblind}
\title{Color palette data for Bokeh Colorblind}
\format{A dataframe with 8 rows and 2 columns:
\describe{
   \item{name}{name of the color}
   \item{value}{hex code for the color}
   }}
\source{
https://bokeh.pydata.org/en/latest/docs/reference/palettes.html
}
\usage{
bokeh_colorblind
}
\description{
This data contains the hex codes from the Bokeh Colorblind palette:
https://bokeh.pydata.org/en/latest/docs/reference/palettes.html
}
\examples{
  bokeh_colorblind
}
\keyword{datasets}
