% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_theme.r
\name{bokeh_colorblind_pal}
\alias{bokeh_colorblind_pal}
\title{Bokeh Colorblind scale}
\usage{
bokeh_colorblind_pal()
}
\description{
This is the Bokeh colorblind scale used for the main colors in the text.
}
\references{
https://bokeh.pydata.org/en/latest/docs/reference/palettes.html
}
\concept{colour bokeh}
