% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_fire.r
\docType{data}
\name{cal_fire}
\alias{cal_fire}
\title{California Wild Fire Data}
\format{
A dataframe with 56,420 rows and 20 columns:
\describe{
   \item{TBD}{TBD}
   }
}
\usage{
cal_fire
}
\description{
To come ...
}
\examples{
  cal_fire
}
\keyword{datasets}
