% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy.r
\docType{data}
\name{happy}
\alias{happy}
\title{Happiness Data}
\format{
A dataframe with 56,420 rows and 20 columns:
\describe{
   \item{TBD}{TBD}
   }
}
\usage{
happy
}
\description{
To come ...
}
\examples{
  happy
}
\keyword{datasets}
