% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_expectancy.r
\docType{data}
\name{life_expectancy}
\alias{life_expectancy}
\title{Life Expectancy Data}
\format{
A dataframe with 56,420 rows and 20 columns:
\describe{
   \item{TBD}{TBD}
   }
}
\usage{
life_expectancy
}
\description{
To come ...
}
\examples{
  life_expectancy
}
\keyword{datasets}
