% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_theme.r
\docType{data}
\name{scale_color_bokehcolorblind}
\alias{scale_color_bokehcolorblind}
\title{rdname scale_bokehcolorblind}
\format{An object of class \code{ScaleDiscrete} (inherits from \code{Scale}, \code{ggproto}, \code{gg}) of length 16.}
\usage{
scale_color_bokehcolorblind
}
\description{
rdname scale_bokehcolorblind
}
\keyword{datasets}
